requirements.txt
torch
soundfile
qwen_tts
torchaudio
# transformers
qwen_tts
transformers==4.57.3
accelerate==1.12.0
huggingface_hub==0.36.2
flash-attention
open-cv
playsound==1.2.2
tkinter
Pillow